requirements.txt
altair==4.2.2
asn1crypto==1.5.1
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
entrypoints==0.4
Faker==18.5.1
favicon==0.7.0
fonttools==4.39.3
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-auth==2.17.3
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
googleapis-common-protos==1.59.0
greenlet==2.0.2
grpcio==1.54.0
grpcio-status==1.48.2
htbuilder==0.6.1
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
lxml==4.9.2
Markdown==3.4.3
markdown-it-py==2.2.0
markdownlit==0.0.7
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
more-itertools==9.1.0
numpy==1.24.3
packaging==23.1
pandas==2.0.3
pandasql==0.7.3
pathlib==1.0.1
pg8000==1.29.8
pickle5==0.0.11
Pillow==9.5.0
plotly==5.14.1
proto-plus==1.22.2
protobuf==3.20.3
psycopg2-binary==2.9.6
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybase64==1.2.3
pydeck==0.8.1b0
Pygments==2.15.1
pymdown-extensions==9.11
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.29.0
rich==13.3.4
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
scramp==1.4.4
seaborn==0.12.2
six==1.16.0
sklearn==0.0
smmap==5.0.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
st-annotated-text==4.0.0
streamlit==1.25.0
streamlit-analytics==0.4.1
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.4
streamlit-embedcode==0.1.2
streamlit-extras==0.2.7
streamlit-faker==0.0.2
streamlit-image-coordinates==0.1.3
streamlit-keyup==0.2.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
tenacity==8.2.2
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.1
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
union==0.1.10
urllib3==1.26.15
validators==0.20.0
zipp==3.15.0
